# docs/enterprise/openapi-enterprise.yaml
openapi: 3.1.0
info:
  title: SCMIRN Enterprise API
  version: 1.0.0
  description: >-
    Prototype API specification. Staff routes are allowlisted only when explicitly
    enabled with production secrets; all enterprise and simulation routes remain disabled
    by the production readiness gate. No government filing or connector exists. This
    document is not evidence of deployment, certification, agency approval, or production readiness.
  x-scmirn-production-status: Prototype only; staff auth is opt-in and operator-configured
servers:
  - url: https://api.scmirn.local
    description: Placeholder hostname; not a deployed service
  - url: http://localhost:5000
    description: Local
...
paths:
...
  /api/v1/triage:
    post:
      tags: [Routing]
      summary: Check for a source-verified civic service route
      description: Requires explicit consent. Returns route guidance only; it never submits a request to an agency.
      x-scmirn-production-status: Allowlisted prototype endpoint; service catalog currently empty
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema: { type: string, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description, consent_to_process]
              properties:
                description: { type: string, minLength: 1, maxLength: 8000 }
                state: { type: string, maxLength: 120 }
                district: { type: string, maxLength: 120 }
                consent_to_process: { type: boolean, enum: [true] }
      responses:
        "200":
          description: Route result with `submission_status` set to `NOT_SUBMITTED` and a null official reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDecision"
        "400": { description: Invalid input or missing consent }
        "409": { description: Idempotency key reused with different input }
        "503": { description: Endpoint disabled during production migration mode }
  /api/v1/jurisdiction/resolve:
    post:
      tags: [Routing]
      summary: Alias for the consented source-gated route check
      description: Returns route guidance only; it never files a government complaint.
      x-scmirn-production-status: Allowlisted prototype endpoint; service catalog currently empty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description, consent_to_process]
              properties:
                description: { type: string, minLength: 1, maxLength: 8000 }
                state: { type: string, maxLength: 120 }
                district: { type: string, maxLength: 120 }
                consent_to_process: { type: boolean, enum: [true] }
      responses:
        "200":
          description: Route result; urgency remains unassessed without a reviewed severity policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDecision"
        "400": { description: Invalid input or missing consent }
        "409": { description: Idempotency key conflict }
...
  /api/v1/services/{service_id}:
    get:
      tags: [Routing]
      summary: Get a registered service's source-verification state
      x-scmirn-production-status: Allowlisted prototype endpoint; no service catalog is seeded
      parameters:
        - in: path
          name: service_id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Service routing policy and evidence requirements }
        "404": { description: Service is not registered }
  /api/v1/evidence/check:
    post:
      tags: [Routing]
      summary: Check evidence identifiers for a registered service
      description: Accepts evidence identifiers only; it does not upload or store evidence files.
      x-scmirn-production-status: Allowlisted prototype endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [service_id]
              properties:
                service_id: { type: string }
                provided_evidence_ids: { type: array, maxItems: 100, items: { type: string } }
      responses:
        "200": { description: Evidence checklist result }
        "400": { description: Invalid request }
  /api/v1/sources/{source_key}:
    get:
      tags: [Routing]
      summary: Get source provenance and verification state
      x-scmirn-production-status: Allowlisted prototype endpoint
      parameters:
        - in: path
          name: source_key
          required: true
          schema: { type: string }
      responses:
        "200": { description: Source metadata, including hash and review state }
        "404": { description: Source not found }
...
components:
  securitySchemes:
    staffSession:
      type: apiKey
      in: cookie
      name: scmirn_staff_session
    staffCsrf:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
    RouteDecision:
      type: object
      required: [decision_id, outcome, issue_type, urgency, urgency_signal, urgent_human_help_required, explanation, route_rule_version, service_registry_version, submission_status, official_reference, sources]
      properties:
        decision_id: { type: string, format: uuid }
        outcome:
          type: string
          enum: [OFFICIAL_HANDOFF_ONLY, ROUTE_UNCERTAIN, JURISDICTION_AMBIGUOUS, URGENT_HUMAN_HELP_REQUIRED]
        issue_type: { type: string }
        urgency:
          type: string
          enum: [LOW, NORMAL, HIGH, CRITICAL, UNASSESSED]
          description: UNASSESSED unless a reviewed, effective severity policy supports a level.
        urgency_signal:
          type: string
          enum: [NONE_DETECTED, POSSIBLE_IMMEDIATE_DANGER]
          description: Phrase signal only; it is not an official severity determination.
        urgent_human_help_required: { type: boolean }
        explanation: { type: string }
        route_rule_version: { type: string }
        service_registry_version: { type: string }
        model_version: { type: [string, "null"] }
        submission_status: { type: string, const: NOT_SUBMITTED }
        official_reference: { type: [string, "null"] }
        sources:
          type: array
          items: { type: object }
...